requirements.txt
python-telegram-bot
SQLAlchemy>=2.0
alembic
pydantic>=2.0
pydantic-settings
pytest
python-dotenv==1.0.1

